+++ tests/Widgets/Scrollbar.cpp
@@ -140,20 +140,20 @@ TEST_CASE("[Scrollbar]")
         REQUIRE(scrollbar->getAutoHide());
     }
 
-    SECTION("VerticalScroll")
+    SECTION("Orientation")
     {
         scrollbar->setSize(100, 20);
-        REQUIRE(!scrollbar->getVerticalScroll());
+        REQUIRE(scrollbar->getOrientation() == tgui::Orientation::Horizontal);
 
         scrollbar->setSize(20, 100);
-        REQUIRE(scrollbar->getVerticalScroll());
+        REQUIRE(scrollbar->getOrientation() == tgui::Orientation::Vertical);
 
         scrollbar->setSize(10, 40);
-        scrollbar->setVerticalScroll(false);
-        REQUIRE(!scrollbar->getVerticalScroll());
+        scrollbar->setOrientation(tgui::Orientation::Horizontal);
+        REQUIRE(scrollbar->getOrientation() == tgui::Orientation::Horizontal);
         REQUIRE(scrollbar->getSize() == tgui::Vector2f(40, 10));
-        scrollbar->setVerticalScroll(true);
-        REQUIRE(scrollbar->getVerticalScroll());
+        scrollbar->setOrientation(tgui::Orientation::Vertical);
+        REQUIRE(scrollbar->getOrientation() == tgui::Orientation::Vertical);
         REQUIRE(scrollbar->getSize() == tgui::Vector2f(10, 40));
     }