+++ changelog.md
@@ -3,6 +3,8 @@ TGUI 1.4  (TBD)
 
 - Added MaxValue getter to Scrollbar
 - Added ScrollbarMaxValue getters to widgets with a scrollbar
+- Added getPixelsPerPoint() to BackendRenderTarget
+- Replace VerticalScroll with Orientation in Slider, Scrollbar and SpinButton
 - Multiple fixes to EditBoxSlider widget